refactor: Add comparison config (+ move the functions there)
It will allow us to use that config next commit
