+++ src/components/comparison/ComparisonEnvironment.vue
@@ -25,77 +25,13 @@
 </template>
 
 <script setup>
-import _ from 'lodash'
-import { ACVImpacts } from '@utils/misc.js'
 import ComparisonTitle from './ComparisonTitle.vue'
 import ComparisonExpandableRow from './ComparisonExpandableRow.vue'
 import ComparisonDefaultCell from './ComparisonDefaultCell.vue'
-
+import { getTotalImpacts, getValuesByImpact } from './comparisonConfig'
 defineProps({
   studies: Array
 })
-
-function getAvailableImpacts(studies) {
-  return studies
-    .reduce((arr, study) => arr.concat(study.acvData?.impacts), [])
-    .filter((item) => !!item)
-    .filter((item) => item.unit === 'Pt')
-}
-
-function getImpacts(studies) {
-  const availableImpacts = getAvailableImpacts(studies)
-  return ACVImpacts.filter((item) =>
-    availableImpacts.map((availableImpact) => availableImpact.name).includes(item.name)
-  )
-}
-
-function getTotalImpacts(study, studies, sumTotalPerYear = false) {
-  const valuesByImpact = getValuesByImpact(study, studies, sumTotalPerYear)
-  if (_.isEmpty(valuesByImpact)) {
-    return null
-  }
-
-  return _.sumBy(Object.values(valuesByImpact))
-}
-
-function getValuesByImpact(study, studies, sumTotalPerYear = false) {
-  const valuesByImpact = {}
-  const impacts = getImpacts(studies)
-  impacts.forEach((impact) => {
-    const impactValue = getImpactValue(impact, study, sumTotalPerYear)
-    if (_.isNull(impactValue)) {
-      return
-    }
-    var label = ACVImpacts.find((item) => item.name == impact.name).label
-    if (label) {
-      valuesByImpact[label] = impactValue
-    } else {
-      valuesByImpact[impact.name] = impactValue
-    }
-  })
-  return valuesByImpact
-}
-
-function getImpactValue(impact, study, sumTotalPerYear = false) {
-  if (!study.acvData) {
-    return null
-  }
-  const valueChains = study.acvData.valuechains
-  let totalPerYear = 0
-  let totalVolume = 0
-  for (const { name, volume } of valueChains) {
-    const totalChainPerYear = study.acvData.impacts
-      .filter((i) => i.name === impact.name)
-      .filter((i) => i.unit === 'Pt')
-      .reduce((arr, item) => arr.concat(item.values), [])
-      .filter((val) => val.valuechain_name === name)
-      .map((val) => val.value * volume)
-      .reduce((s, item) => s + item, 0)
-    totalVolume += volume
-    totalPerYear += totalChainPerYear
-  }
-  return sumTotalPerYear ? totalPerYear : totalPerYear / totalVolume
-}
 </script>
 
 <style scoped lang="scss">
